Fix constructor type signature
All the subclasses take a configuration dictionary as constructor
argument, so the interface definition should also mention it. In
addition to improving documentation, this change will also make it safer
to use reflection to look up and instantiate specific implementations of
the interface since the constructor signature now is part of the
official interface.