Merge layers
In docker the filesystem at the end of each RUN statement gets
serialized into the image. This means that if we do apk add in one RUN
and apk del in another one, the image size doesn't actually reduce since
all the additions were serialized into a layer after the first step.